consolidated the image and motion generators into the ImageInterface component
