+++ src/components/imageInterface/ImageInterface.jsx
@@ -2,10 +2,83 @@ import React, { useState } from "react";
 import { NavLink } from "react-router-dom";
 import { TextField, Button, Grid } from "@mui/material";
 import ArrowForwardOutlinedIcon from "@mui/icons-material/ArrowForwardOutlined";
-import Header from "../header/Heading";
+import Header from "../header/header";
 import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
+// import ImageGenerator from '../ImageGenerator'
+import axios from '../axiosConfig/axiosConfig'
+import Spinner from '../Spinner/Spinner'
+import ReactPlayer from 'react-player';
+
+
+let ImgId = '';
 
 function ImageInterface() {
+  const [prompt, setPrompt] = useState('');
+  const [imageUrl, setImageUrl] = useState('');
+  const [loading, setLoading] = useState(false);
+  const [motionUrl, setMotionUrl] = useState('');
+
+  const generateImage = async () => {
+    try {
+      setLoading(true);
+      const response = await axios.post('/generations', {
+        prompt: prompt,
+        height: 512,
+        modelId: '',
+        width: 512,
+        alchemy: true,
+        photoReal: true,
+        photoRealStrength: 0.5,
+        presetStyle: "CINEMATIC"
+      });
+      console.log(response);
+      const generationId = response.data.sdGenerationJob.generationId;
+      console.log(generationId);
+      setTimeout(async () => {
+        try {
+          const imageResponse = await axios.get(`/generations/${generationId}`);
+          const generatedImages = imageResponse.data.generations_by_pk.generated_images[0].url;
+          ImgId = imageResponse.data.generations_by_pk.generated_images[0].id;
+          console.log(generatedImages);
+          console.log(ImgId);
+          setImageUrl(generatedImages);
+
+        } catch (error) {
+          console.error('Error fetching image URLs:', error);
+        } finally {
+          setLoading(false);
+        }
+      }, 28000);
+    } catch (error) {
+      console.error('Error generating image:', error);
+    }
+  };
+
+  const generateMotion = async () => {
+    try {
+      setLoading(true);
+      const motionResponse = await axios.post('/generations-motion-svd', {
+        imageId: ImgId,
+        motionStrength: 5,
+      });
+
+      const motionGenerationId = motionResponse.data.motionSvdGenerationJob.generationId;
+
+      setTimeout(async () => {
+        try {
+          const motionImageResponse = await axios.get(`/generations/${motionGenerationId}`);
+          const generatedMotion = motionImageResponse.data.generations_by_pk.generated_images[0].motionMP4URL;
+          setMotionUrl(generatedMotion);
+        } catch (error) {
+          console.error('Error fetching motion URLs:', error);
+        } finally {
+          setLoading(false);
+        }
+      }, 80000); // Adjust the delay as needed
+    } catch (error) {
+      console.error('Error generating motion:', error);
+    }
+  };
   return (
     <Grid container direction="column" marginTop={2} paddingX={2}>
       <div style={{ position: "relative" }}>
@@ -46,6 +119,8 @@ function ImageInterface() {
               }}
               variant="outlined"
               placeholder="Enter a prompt to generate..."
+              value={prompt}
+              onChange={(e) => setPrompt(e.target.value)}
             />
           </Grid>
           <Grid item xs={1}>
@@ -59,12 +134,15 @@ function ImageInterface() {
               }}
               fullWidth
               variant="contained"
+              onClick={generateImage}
             >
               <ArrowForwardOutlinedIcon />
             </Button>
+
           </Grid>
         </Grid>
       </div>
+      
       <Grid container justifyContent="space-between" gap={2}>
         <Grid
           item
@@ -76,7 +154,20 @@ function ImageInterface() {
           height="90vh"
           borderRadius={1}
           padding={2}
-        ></Grid>
+        >{loading ? <Spinner /> : imageUrl && <img src={imageUrl} alt="Generated Image" />}<Button
+          sx={{
+            bgcolor: "rgb(231, 132, 48)",
+            height: "50px",
+            "&:hover": {
+              bgcolor: "rgb(231, 132, 48)",
+            },
+          }}
+          fullWidth
+          variant="contained"
+          onClick={generateMotion}
+        >
+            <ArrowForwardOutlinedIcon />
+          </Button></Grid>
         <Grid
           item
           sm={12}
@@ -87,10 +178,10 @@ function ImageInterface() {
           height="90vh"
           borderRadius={1}
           padding={2}
-        ></Grid>
+        >{loading ? <Spinner /> : motionUrl && <ReactPlayer playing muted={true} loop={true} url={motionUrl} />}</Grid>
       </Grid>
     </Grid>
   );
-}
 
+}
 export default ImageInterface;
